scripts clean up

+++ replicate_ppopp20.sh
@@ -1,19 +1,14 @@
 #!/bin/bash
 
-### Downloading dataset (If you have download the dataset, this part can be commented out)
-mkdir ./data
-wget -O ./data/Points.zip https://www.dropbox.com/sh/ab7f8gut3nh22ym/AAA0QXrC3kS0L4iHS2T0kpg-a?dl=0
-unzip ./data/Points.zip -d ./data/
-rm -f ./data/Points.zip
-
-
-### Intalling MatRox (It can be commented out after installing MatRox)
-mkdir build
-cd build
-cmake ..
-cd sympiler
-make
-
+### Downloading dataset (If the data folder and 
+### sunflower8w80000.points.bin are created, it skips downloading the dataset.)
+if [ ! -d "data/" ] && [! -f "data/sunflower8w80000.points.bin"]; then
+ mkdir ./data
+ echo "Downloading point sets ..."
+ wget -O ./data/Points.zip https://www.dropbox.com/sh/ab7f8gut3nh22ym/AAA0QXrC3kS0L4iHS2T0kpg-a?dl=0
+ unzip ./data/Points.zip -d ./data/
+ rm -f ./data/Points.zip
+fi
 
 ### Run the jobs on comet or knl
 cometORknl=1  # Comet is 1, KNL is 2
@@ -21,11 +16,12 @@ if [ "$#" -eq 1 ]; then
 cometORknl=$1
 fi
 
-cd ../../
 
 if [ $cometORknl -eq 1 ]; then
+	echo "Running on Comet ... "
 	sbatch matrox_comet.sh 
 else
+	echo "running on KNL ..."
 	sbatch matrox_knl.sh
 fi
